A version that implements data processing using multiprocsses within lasif instead of using the built in function in pyasdf.
Not quite happy with this version, yet as it is still not faster.

To be fast it will be important that threads can write the results
to a file as soon as they are done.

Furthermore, the bottleneck really isn't the pyasdf code itself,
but the instrument response removal function in obspy.

To get significant improvements, we should look into improving these
functions within obspy itself.